refactor commands into receivable ones and sendable ones or both. Also made the tracker connection and it can actually connect to a fake tracker

+++ peer/src/main/java/fr/ystat/tracker/criterions/FilesizeCriterion.java
@@ -1,10 +1,10 @@
 package fr.ystat.tracker.criterions;
 
 public class FilesizeCriterion implements ICriterion{
-	private final int fileSize;
+	private final long fileSize;
 	private final ComparisonType comparisonType;
 
-	public FilesizeCriterion(int fileSize, ComparisonType comparisonType) {
+	public FilesizeCriterion(long fileSize, ComparisonType comparisonType) {
 		this.fileSize = fileSize;
 		this.comparisonType = comparisonType;
 	}
